Fix for Issue #3 checking bpy.app.build_revision which wasn't an integer.  This can all be removed once the next beta is out.
Bumped revision to 0.8.1.

Added a DOS batch file for creating a ZIP archive.  Blender can install ZIP archives itself, plus this works around the problem of not being able to put a version number in the io_smd_tools.py filename.

Might want to put a README file in there as well, although it will be extracted into the addons directory. README-io_smd_tools.txt with install instructions and a link to the Google Code homepage perhaps?

To use the batch file, download 7-ZIP from:
http://www.7-zip.org/download.html
and put 7za.exe into your blender SVN directory.  To make a new release, just update the revision REV in the BATCH file and then double-click it, and it will create a zip archive for uploading.